requirements.txt
openai
whisper-openai
gradio
numpy
torch
librosa
pydub
ffmpeg-python
soundfile
pytest
black
pydantic 